Fix logic for the GCD.

+++ Support/DataGenerator/simc_copied.hpp
@@ -6,8 +6,8 @@
 
 namespace simc_copied
 {
-    std::string className_from_classID(size_t classID);
-    std::string specName_from_specID(size_t specID);
+    std::string className_from_classID(int classID);
+    std::string specName_from_specID(int specID);
 
     int translate_spec_str(int ptype, const std::string& spec_str);
     std::string specialization_string(int spec);
