+++ Support/DataGenerator/simc_data.hpp
@@ -10,28 +10,28 @@ namespace simc_data
 {
     //////////////////////////////////////////////////
     // common
-    size_t specIdx_from_specID(size_t specID);
+    int specIdx_from_specID(int specID);
 
-    std::vector<size_t> slug_to_itemIDs(const std::string& slug);
-    std::vector<size_t> slug_to_spellIDs(const std::string& slug);
+    std::vector<int> slug_to_itemIDs(const std::string& slug);
+    std::vector<int> slug_to_spellIDs(const std::string& slug);
 
-    std::set<size_t> specIDs_from_classID(size_t classID);
-    size_t classID_from_specID(size_t specID);
+    std::set<int> specIDs_from_classID(int classID);
+    int classID_from_specID(int specID);
 
     //////////////////////////////////////////////////
     // talents
 
-    std::set<size_t> talentIDs_from_specID(size_t specID);
+    std::set<int> talentIDs_from_specID(int specID);
 
     struct talent_t
     {
-        size_t id;
+        int id;
         const char* name;
-        size_t row;
-        size_t col;
-        size_t specID;
+        int row;
+        int col;
+        int specID;
     };
-    talent_t talent_info(size_t talentID, bool throwIfNotFound = true);
+    talent_t talent_info(int talentID, bool throwIfNotFound = true);
 
     namespace detail
     {
@@ -41,33 +41,33 @@ namespace simc_data
         };
     } // namespace detail
     using talent_set_t = std::set<talent_t, detail::talent_comparator>;
-    talent_set_t talents_from_specID(size_t specID);
+    talent_set_t talents_from_specID(int specID);
 
     //////////////////////////////////////////////////
     // items
     struct item_t
     {
-        size_t id;
+        int id;
         const char* name;
     };
-    item_t item_info(size_t itemID, bool throwIfNotFound = true);
+    item_t item_info(int itemID, bool throwIfNotFound = true);
 
-    std::map<std::string, std::set<size_t>> itemsets_from_classID(size_t classID);
+    std::map<std::string, std::set<int>> itemsets_from_classID(int classID);
 
     //////////////////////////////////////////////////
     // artifacts
-    size_t artifactID_from_specID(size_t specID);
-    std::set<size_t> artifactTraitIDs_from_artifactID(size_t artifactID);
-    std::pair<size_t, size_t> artifactItemIDs_from_artifactID(size_t artifactID);
-    std::pair<size_t, size_t> artifactItemIDs_from_specID(size_t specID);
+    int artifactID_from_specID(int specID);
+    std::set<int> artifactTraitIDs_from_artifactID(int artifactID);
+    std::pair<int, int> artifactItemIDs_from_artifactID(int artifactID);
+    std::pair<int, int> artifactItemIDs_from_specID(int specID);
 
     struct artifact_trait_t
     {
-        size_t id;
+        int id;
         const char* name;
-        size_t max_rank;
+        int max_rank;
     };
-    artifact_trait_t artifact_trait_info(size_t artifactTraitID, bool throwIfNotFound = true);
+    artifact_trait_t artifact_trait_info(int artifactTraitID, bool throwIfNotFound = true);
 
     namespace detail
     {
@@ -82,15 +82,15 @@ namespace simc_data
         };
     } // namespace detail
     using artifact_trait_set_t = std::set<artifact_trait_t, detail::artifact_trait_comparator>;
-    artifact_trait_set_t artifactTraits_from_artifactID(size_t artifactID);
+    artifact_trait_set_t artifactTraits_from_artifactID(int artifactID);
 
     //////////////////////////////////////////////////
     // spells
-    std::set<size_t> spellIDs_from_classID(size_t classID);
+    std::set<int> spellIDs_from_classID(int classID);
 
     struct spell_t
     {
-        size_t id;
+        int id;
         const char* name;
         bool is_ability;
         bool is_passive;
@@ -98,15 +98,15 @@ namespace simc_data
         float gcd;
         float duration;
         float cooldown;
-        size_t charges;
+        int charges;
         float charge_cooldown;
-        size_t max_stack;
+        int max_stack;
         float min_range;
         float max_range;
         const char* description;
         const char* tooltip;
     };
-    spell_t spell_info(size_t spellID, bool throwIfNotFound = true);
+    spell_t spell_info(int spellID, bool throwIfNotFound = true);
 
     namespace detail
     {
@@ -121,25 +121,25 @@ namespace simc_data
         };
     } // namespace detail
     using spell_set_t = std::set<spell_t, detail::spell_comparator>;
-    spell_set_t spells_from_classID(size_t classID);
+    spell_set_t spells_from_classID(int classID);
 
     //////////////////////////////////////////////////
     // spell effects
-    std::set<size_t> spellEffectIDs_from_spellID(size_t spellID);
+    std::set<int> spellEffectIDs_from_spellID(int spellID);
 
     struct spelleffect_t
     {
-        size_t id;
-        size_t index;
-        size_t type;
-        size_t subtype;
-        size_t trigger_spell_id;
+        int id;
+        int index;
+        int type;
+        int subtype;
+        int trigger_spell_id;
         int val1;
         int val2;
         int val3;
         int die_sides;
     };
-    spelleffect_t spelleffect_info(size_t spellEffectID, bool throwIfNotFound = true);
+    spelleffect_t spelleffect_info(int spellEffectID, bool throwIfNotFound = true);
 
     namespace detail
     {
@@ -149,5 +149,5 @@ namespace simc_data
         };
     } // namespace detail
     using spelleffect_set_t = std::set<spelleffect_t, detail::spelleffect_comparator>;
-    spelleffect_set_t spellEffects_from_spellID(size_t spellID);
+    spelleffect_set_t spellEffects_from_spellID(int spellID);
 } // namespace simc_data
